requirements.txt
python-dotenv==1.0.1

langchain==0.1.12
langchain-openai==0.0.6
langchain-community==0.0.29
langchain-text-splitters==0.0.1
langchain-utils==0.6.0

streamlit==1.31.0
# streamlit-cookies-manager==0.2.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2

trycourier==4.4.0

altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
chromadb==0.4.24
